+++ src_c/klua/klua_net/klua_ktcp.c
@@ -557,10 +557,10 @@ static void check_param_klua_ktcp(lua_State* L, int from, klua_ktcp_param_t* p_p
         switch (idx)
         {
         case 0:
-            p_param->rbuf_max = luaL_checkinteger(L, ud);
+            p_param->rbuf_max = (int)luaL_checkinteger(L, ud);
             break;
         case 1:
-            p_param->read_max = luaL_checkinteger(L, ud);
+            p_param->read_max = (int)luaL_checkinteger(L, ud);
             break;
         case 2:
             p_param->tls = luaL_checkboolean(L, ud);
